requirements.txt
pytesseract
pdf2image
opencv-python
pillow
streamlit
